apply user list popover
